pyproject.toml
[tool.poetry]
name = "it-solution-parsing-api"
version = "0.1.0"
description = "Тестовое задание на позицию Backend-разработчик на языке Python"
authors = ["Дмитрий Панин <dimitri-05@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"

[tool.poetry.group.parse.dependencies]
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"

[tool.poetry.group.database.dependencies]
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.1"

[tool.poetry.group.security.dependencies]
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"
black = "^24.4.2"


[tool.isort]
profile = "black"
line_length = 100
extend_skip = ["migrations"]
skip_gitignore = true

[tool.black]
line-length = 100
extend-exclude = "migrations"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
